Add configuration file, use pathlib
Review for UBC-ECE-Sasha/PIM-compression#9

Config:

Rather than hardcoding configuration in the script, using a
configuration file would allow for easier changes.

It would also allow ignoring of the real configuration file,
so that multiple people could use different configuration files.
An example of a configuration file could contain what is usually
used.

Paths:

Using raw string paths is fraught with problems. In the
current implementation if one uses the path 'out' as the directory,
a failure occurs since the directory is just appended to
the file name.

pathlib allows paths to be treated as objects, and correctly joined.

Opening files:

Currently open files are not properly closed. Use a context
manager to correctly manage the resource and close it when the
resource manager block has exited.

Division by zero:

Currently if the files are missing that are in the 'files' dict,
division by zero occurs.

Emit a warning that no files were found, rather than an exception
occurring.